Splitt opp redigerbar sjekken for fastestting av inntekt
