+++ apps/etterlatte-saksbehandling-ui/client/src/components/etteroppgjoer/forbehandling/EtteroppgjoerForbehandlingOversikt.tsx
@@ -7,22 +7,32 @@ import { Link } from 'react-router-dom'
 import { EtteroppjoerForbehandlingSteg } from '~components/etteroppgjoer/forbehandling/stegmeny/EtteroppjoerForbehandlingStegmeny'
 import { ResultatAvForbehandling } from '~components/etteroppgjoer/components/resultatAvForbehandling/ResultatAvForbehandling'
 import { BrevutfallAvForbehandling } from '~components/etteroppgjoer/components/resultatAvForbehandling/BrevutfallAvForbehandling'
+import { EtteroppgjoerBehandlingStatus } from '~shared/types/Etteroppgjoer'
+import { enhetErSkrivbar } from '~components/behandling/felles/utils'
+import { useInnloggetSaksbehandler } from '~components/behandling/useInnloggetSaksbehandler'
 
 export const EtteroppgjoerForbehandlingOversikt = () => {
-  const etteroppgjoer = useEtteroppgjoer()
+  const innloggetSaksbehandler = useInnloggetSaksbehandler()
+
+  const { beregnetEtteroppgjoerResultat, behandling } = useEtteroppgjoer()
+
+  const erRedigerbar =
+    (behandling.status == EtteroppgjoerBehandlingStatus.OPPRETTET ||
+      behandling.status == EtteroppgjoerBehandlingStatus.BEREGNET) &&
+    enhetErSkrivbar(behandling.sak.enhet, innloggetSaksbehandler.skriveEnheter)
 
   return (
     <VStack gap="10" paddingInline="16" paddingBlock="16 4">
       <Heading size="xlarge" level="1">
-        Etteroppgjør for {etteroppgjoer.behandling.aar}
+        Etteroppgjør for {behandling.aar}
       </Heading>
       <BodyShort>
-        <b>Skatteoppgjør mottatt:</b> {formaterDato(etteroppgjoer.behandling.opprettet)}
+        <b>Skatteoppgjør mottatt:</b> {formaterDato(behandling.opprettet)}
       </BodyShort>
       <Inntektsopplysninger />
-      <FastsettFaktiskInntekt />
+      <FastsettFaktiskInntekt erRedigerbar={erRedigerbar} />
 
-      {!!etteroppgjoer.beregnetEtteroppgjoerResultat && (
+      {!!beregnetEtteroppgjoerResultat && (
         <VStack gap="4">
           <ResultatAvForbehandling />
           <BrevutfallAvForbehandling />
@@ -32,10 +42,7 @@ export const EtteroppgjoerForbehandlingOversikt = () => {
       <Box borderWidth="1 0 0 0" borderColor="border-subtle" paddingBlock="8 16">
         <HStack width="100%" justify="center">
           <div>
-            <Button
-              as={Link}
-              to={`/etteroppgjoer/${etteroppgjoer.behandling.id}/${EtteroppjoerForbehandlingSteg.BREV}`}
-            >
+            <Button as={Link} to={`/etteroppgjoer/${behandling.id}/${EtteroppjoerForbehandlingSteg.BREV}`}>
               Gå til brev
             </Button>
           </div>
